requirements.txt
numpy==1.15.2
opencv_python==3.4.3.18
scikit_image==0.14.1
Keras==2.2.4
Flask==1.0.2
Pillow==5.3.0
skimage==0.0
